cstPiecewise {CBCtool}	R Documentation

Categorizes continuous predictors using piecewise constant functions.

Description

Categorizes continuous predictors using piecewise constant functions.

Usage


cstPiecewise(data, quantils = NULL, QUALI = NULL)

Arguments

quantils
Quantiles used to categorize continous variables

QUALI
Columns of X corresponding to qualitative variables

X
Data frame of predictors

Value

An object of class "cstPiecewise" which contains the list of the piecewise function of each variable and the X matrix produced changing the quantitative variables by piecewise functions.

Examples


cstPiecewise(CBCsimulatedData[,1:20],QUALI = c(21,22))

[Package CBCtool version 0.0.0.9000 Index]