piecewiseLasso {CBCtool}	R Documentation

Categorizes continuous predictors using piecewise constant functions, runs a Lasso regression and calculates selection probabilities

Description

Categorizes continuous predictors using piecewise constant functions, runs a Lasso regression and calculates selection probabilities

Usage


piecewiseLasso(
  X,
  y,
  QUALI = NULL,
  quantils = NULL,
  foldid = NULL,
  nfold = 20,
  weight = 1,
  selectProba = TRUE,
  verbose = TRUE,
  ...
)

Arguments

X
Data frame of p predictors

y
Binary response variable

QUALI
Columns of X corresponding to qualitative variables

quantils
Quantiles used to categorize continous variables

foldid
Optional vector of values between 1 and nfold identifying what fold each observation is in. If supplied, nfold can be missing.

nfold
Number of folds to be used in cross-validation - default is 20. Smallest value allowed is nfolds=3

weight
Positive class weigth. Default is 1.

selectProba
If TRUE selection probabilities are estimated

verbose
Shows starting and ending times. Default is TRUE

Others
parameters of the cv.glmnet function (parallel …)

Value

An object of class "piecewiseLasso" with the following components:

An object of class "cv.glmnet"

An object of class "cstPiecewise" which contains the list of the piecewise function of each variable and the X matrix produced changing the quantitative variables by piecewise functions.

The p-vector gathering the selection probabilities of each variable IF parameter selectProba is TRUE

Examples


pl <- piecewiseLasso(CBCsimulatedData[1:22],CBCsimulatedData[23],QUALI = c(21,22))

[Package CBCtool version 0.0.0.9000 Index]