plotPiecewiseLasso {CBCtool}	R Documentation

Plot the Lasso estimates in terms of variation of OR and selection probabilities, represented by a light to dark gradient of gray (the darker, the more stable the results are).

Description

Plot the Lasso estimates in terms of variation of OR and selection probabilities, represented by a light to dark gradient of gray (the darker, the more stable the results are).

Usage


plotPiecewiseLasso(
  piecewiseLasso,
  var = NULL,
  cons = TRUE,
  refThresholds = NULL,
  distrubution = FALSE,
  X = NULL,
  y = NULL
)

Arguments

piecewiseLasso
Fitted "piecewiseLasso" model

var
Variables to plot

cons
If TRUE, the recommended thresholds will be plot.

refThresholds
Particulars thresholds the user wants plot

X
Data frame of predictors (needed to plot the distribution)

y
Binary response variable (needed to plot the distribution)

distribution
If TRUE the distribution of the observations will be plot

Examples


pl<-piecewiseLasso(CBCsimulatedData[1:22],CBCsimulatedData[23],QUALI = c(21,22))
plotPiecewiseLasso(pl,distrubution = TRUE,X=CBCsimulatedData[1:20], y= CBCsimulatedData[,23])


[Package CBCtool version 0.0.0.9000 Index]