simulation {CBCtool}	R Documentation

Simulates CBC data from mean vector and covariance matrix

Description

Simulates CBC data from mean vector and covariance matrix

Usage


simulation(
  mean0 = mean0,
  mean1 = mean1,
  cov0 = cov0,
  cov1 = cov1,
  QUALI = NULL,
  n = 400,
  propotion = 0.07,
  replaceNeg = TRUE,
  minValue = NULL
)

Arguments

mean0
mean p-vector of the negatives, with p=22 predictors

mean1
mean p-vector of the positives, with p=22 predictors

cov0
covariance pxp-matrix of the negatives

cov1
covariance pxp-matrix of the positives

QUALI
Columns of X corresponding to qualitative variables

n
number of observations to be generated

replaceNeg
If TRUE the simulated negative values will be replace either by zero or by a minimal value if the argument minValue is given -> If TRUE the simulated negative values will be replaced either by zero or by a minimal value, if the argument minValue is given

minValue
optional vector of the p minimal values used to replace the simulated negative values -> p-vector of minimal values (X values are expected to be >=0)

proportion
proportion of the minority class (positives)

Value

data frame of the simulated data

Examples


CBCsimulatedData<-simulation(mean0,mean1,cov0,cov1,QUALI= c(21,22))

[Package CBCtool version 0.0.0.9000 Index]