weightSettings {CBCtool}	R Documentation

Graphical help to choose the weighting of the classes Performs a k-fold cross-validation with several weighting using the piecewiseLasso function and then produces a plot showing recall and precision for each weigthing.

Description

Graphical help to choose the weighting of the classes Performs a k-fold cross-validation with several weighting using the piecewiseLasso function and then produces a plot showing recall and precision for each weigthing.

Usage


weightSettings(
  X,
  y,
  QUALI = NULL,
  ngroups = 20,
  sample = 1,
  Posweights = c(1, 5, 10, 20, 30, 50),
  recallRef = 0.85,
  precisionRef = NULL,
  aucRef = NULL,
  median = FALSE,
  quantils = NULL,
  foldid = NULL,
  nfold = 20,
  ...
)

Arguments

X
Data frame of p predictors

y
Binary response variable

QUALI
Columns of X corresponding to qualitative variables

ngroups
Number of folds to be used in the k-fold cross-validation. Default is 20.

sample
portion of the data to use

Posweights
list of weight values to evaluate. Default is c(1, 5, 10, 20,30,50).

recallRef
horizontal line to be plotted as reference value for recall. Default is 0.85

precisionRef
horizontal line to be plotted as reference value for precision

aucRef
horizontal line to be plotted as reference value for AUC

median
if TRUE median values of recall, precision and AUC are plotted

Others
parameters of the piecewiseLasso function (quantils,nfold,parallel …)

Examples


weightSettings(CBCsimulatedData[,1:20],CBCsimulatedData[,23],QUALI = c(21,22),sample = 1)

[Package CBCtool version 0.0.0.9000 Index]